Move cross account log subscriptions out of main file
These are now set in terraform.tfvars file locally.
An example of this has been uploaded to SSM Parameter store in Shared
services. It can be found under `/staff-device/shared-services/terraform.tfvars`

Ensure logs are sent from shared services to all 3 accounts so they are
identical and alarms can be configured in the same way.